requirements.txt
streamlit==1.36.0
opencv-python-headless==4.8.1.78
Pillow==10.2.0
numpy==1.26.4
Jinja2==3.1.3
jsonschema==4.21.1
pytesseract==0.3.10
typing_extensions==4.10.0
